% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ncdc_station_data.R
\name{get_ncdc_station_data}
\alias{get_ncdc_station_data}
\title{Get met station data from NCDC}
\usage{
get_ncdc_station_data(station_id, startyear, endyear)
}
\arguments{
\item{station_id}{a station identifier composed of the station's USAF and WBAN numbers, separated by a hyphen.}

\item{startyear}{the starting year for the collected data.}

\item{endyear}{the ending year for the collected data.}
}
\description{
Obtain one or more years of meteorological data for a station from the NCDC hourly global meteorology archive.
}

