% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_tz_offset.R
\name{get_tz_offset}
\alias{get_tz_offset}
\title{Get time zone offset from GMT}
\usage{
get_tz_offset(x, y, proj4str = "")
}
\arguments{
\item{x}{coordinate in the x direction.}

\item{y}{coordinate in the y direction.}

\item{proj4str}{an optional PROJ.4 string for the provided coordinates;
not providing a value assumes lat/lon coordinates.}
}
\description{
Obtain the time zone difference in hours between GMT and a
provided location.
}

