% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_data.R
\name{get_met_data}
\alias{get_met_data}
\title{Get data from a meteorological station}
\usage{
get_met_data(station_id, years = NULL, full_data = FALSE,
  add_fields = NULL, use_local_files = FALSE, local_file_dir = NULL)
}
\arguments{
\item{station_id}{A station identifier composed of the station's USAF and
WBAN numbers, separated by a hyphen.}

\item{years}{The years for which station met data will be collected. If not
specified then all records for all available years will be obtained for the
station.}

\item{full_data}{Include additional meteorological data found in the
dataset's additional data section?}

\item{add_fields}{A vector of categories for additional meteorological data
to include (instead of all available categories).}

\item{use_local_files}{Option to use data files already available locally.}

\item{local_file_dir}{Path to local meteorological data files. If specified,
then data files will be obtained from this location and not from the remote
data store.}
}
\value{
Returns a tibble with 18 variables. Times are recorded
using the Universal Time Code (UTC) in the source data. Times are adjusted
to local standard time for the station's locale.
\describe{
\item{id}{A character string identifying the fixed weather station
from the USAF Master Station Catalog identifier and the WBAN identifier.}
\item{time}{A datetime value representing the observation time.}
\item{wd}{The angle of wind direction, measured in a clockwise direction,
between true north and the direction from which the wind is blowing. For
example, \code{wd = 90} indicates the wind is blowing from due east. \code{wd = 225}
indicates the wind is blowing from the south west. The minimum value is \code{1},
and the maximum value is \code{360}.} \item{ws}{Wind speed in meters per second.
Wind speed in feet per second can be estimated by \code{ws * 3.28084}.}
\item{ceil_hgt}{The height above ground level of the lowest clould cover or
other obscuring phenomena amounting to at least 5/8 sky coverate. Measured in
meters. Unlimited height (no obstruction) is denoted by the value \code{22000}.}
\item{temp}{Air temperature measured in degrees Celsius. Conversions to
degrees Farenheit may be calculated with \code{(temp * 9) / 5 + 32}.}
\item{dew_point}{The temperature in degrees Celsius to which a given parcel
of air must be cooled at constant pressure and water vapor content in order
for saturation to occur.} \item{atmos_pres}{The air pressure in hectopascals
relative to Mean Sea Level (MSL).} \item{rh}{Relative humidity, measured as a
percentage, as calculated using the August-Roche-Magnus approximation.}
}
}
\description{
Obtain one or more years of meteorological data for a particular station.
}
\examples{
\dontrun{
# Obtain two years of data from
# data files stored on disk (in this
# case, inside the package itself)
met_data_local <- 
  get_met_data(
    station_id = "999999-63897",
    years = 2013:2014,
    use_local_files = TRUE,
    local_file_dir = system.file(package = "stationary")
  )
}

}
